+++ sky-take-out/sky-server/src/main/java/com/sky/service/AddressBookService.java
@@ -31,7 +31,7 @@ public interface AddressBookService {
      * @param id 地址id
      * @return 地址
      */
-    AddressBook getById(Integer id);
+    AddressBook getById(Long id);
 
     /**
      * 修改地址
